lab2/requirements.txt
numpy
argparse
gym
matplotlib
Pillow
scipy
gym_minigrid
torchvision